Update README.md
Just adding the README.md for now and changed to branch master. The description is not much, but as of right now it is all that is needed for what I need to do. I will update it all later after I have finished with everything and setting up all that I have to set up. I will work on doing all that after I Finish the rest of the project.